+++ package.json
@@ -10,13 +10,13 @@
   "license": "MIT",
   "scripts": {
     "clean": "rimraf dist",
-    "lint": "eslint --color ./bitmovin ./tests ./tests_it ./examples",
+    "lint": "eslint --color ./bitmovin ./tests ./examples",
     "lint-src": "eslint --color ./bitmovin",
     "test": "node scripts/testRunner.js ./tests",
     "coverage": "node scripts/testRunner.js ./tests --coverage",
     "it-coverage": "rm -rf ./coverage && npm run analyze-coverage && npm run check-coverage",
-    "format": "prettier --config ./.prettierrc --write \"./{bitmovin,examples,tests,tests_it}/**/*.js\"",
-    "format-check": "prettier --config ./.prettierrc --list-different \"./{bitmovin,examples,tests,tests_it}/**/*.js\"",
+    "format": "prettier --config ./.prettierrc --write \"./{bitmovin,examples,tests}/**/*.js\"",
+    "format-check": "prettier --config ./.prettierrc --list-different \"./{bitmovin,examples,tests}/**/*.js\"",
     "build": "yarn clean && yarn build:standalone:dev && yarn build:standalone:prod && yarn build:modules && yarn build:modules-types",
     "build:standalone:dev": "NODE_ENV=development webpack --config=./webpack4.config.js",
     "build:standalone:prod": "NODE_ENV=production webpack --config=./webpack4.config.js",
